refactor: split server index into route modules
